changed medication strengths to being a double instead of an int

+++ app/femr/data/models/core/IMedicationGenericStrength.java
@@ -30,9 +30,6 @@ public interface IMedicationGenericStrength {
 
     void setConceptMedicationUnit(IConceptMedicationUnit conceptMedicationUnit);
 
-
-    //TODO: how to handle these when different implementations have different names for the method
-
     IMedicationGeneric getMedicationGeneric();
 
     void setMedicationGeneric(IMedicationGeneric medicationGeneric);
@@ -41,7 +38,7 @@ public interface IMedicationGenericStrength {
 
     void setDenominator(boolean isDenominator);
 
-    int getValue();
+    Double getValue();
 
-    void setValue(int value);
+    void setValue(Double value);
 }
