+++ app/femr/data/models/mysql/MedicationGenericStrength.java
@@ -21,6 +21,7 @@ package femr.data.models.mysql;
 import femr.data.models.core.IMedicationGenericStrength;
 import femr.data.models.core.IMedicationGeneric;
 import femr.data.models.core.IConceptMedicationUnit;
+import femr.data.models.mysql.concepts.ConceptMedicationUnit;
 
 import javax.persistence.*;
 
@@ -39,7 +40,7 @@ public class MedicationGenericStrength implements IMedicationGenericStrength {
     @Column(name = "isDenominator", nullable = false)
     private boolean isDenominator;
     @Column(name = "value", unique = true, nullable = false)
-    private int value;
+    private Double value;
 
     @Override
     public int getId() {
@@ -74,12 +75,12 @@ public class MedicationGenericStrength implements IMedicationGenericStrength {
     }
 
     @Override
-    public int getValue() {
+    public Double getValue() {
         return value;
     }
 
     @Override
-    public void setValue(int value) {
+    public void setValue(Double value) {
         this.value = value;
     }
 }
